qr_CodeExtractor/requirements.txt
boto3==1.34.20
pyzbar==0.1.9
Pillow==10.2.0
opencv-python-headless
pyzbar[scripts]